+++ packages/mobile/example/lib/component/navigation/app_bar_example.dart
@@ -187,6 +187,13 @@ class _AppBar6 extends StatelessWidget {
           foregroundColor: Colors.white,
           centerTitle: true,
           statusBarBrightness: Brightness.dark,
+          shadows: [
+            BoxShadow(
+              color: Colors.red,
+              blurRadius: 20,
+              spreadRadius: 4,
+            )
+          ],
         ),
         actions: [
           IconButton(