+++ packages/mobile/lib/components/timepicker/zero_time_picker.dart
@@ -5,7 +5,6 @@ import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
-import 'package:zog_ui/styles/component/timepicker_style.dart';
 import 'package:zog_ui/zog_ui.dart';
 
 const Duration _kDialogSizeAnimationDuration = Duration(milliseconds: 200);