+++ packages/mobile/lib/styles/component/textfield_style.dart
@@ -345,13 +345,16 @@ class ZeroTextfieldStyle {
 
   ZeroTextfieldStyle merge(ZeroTextfieldStyle? other) {
     if (other == null) return this;
+
     return ZeroTextfieldStyle(
-        inputDecorationType: other.inputDecorationType,
-        textfieldSize: other.textfieldSize,
-        enabled: other.enabled,
-        error: other.error,
-        focusedBorderColor: other.focusedBorderColor,
-        focusedColor: other.focusedColor);
+      inputDecorationType: other.inputDecorationType,
+      textfieldSize: other.textfieldSize,
+      enabled: other.enabled,
+      error: other.error,
+      focusedBorderColor: other.focusedBorderColor,
+      focusedColor: other.focusedColor,
+      fillColor: other.fillColor,
+    );
   }
 
   InputDecorationTheme toInputDecorationTheme() {